+++ CurrencySpender/Helpers/VersionHelper.cs
@@ -37,8 +37,12 @@ namespace CurrencySpender.Helpers
                 C.SelectedCurrencies.Add(37549);
                 C.SelectedCurrencies.Add(37550);
             }
+            if (LowerVersionThan("1.2.2"))
+            {
+                C.SelectedCurrencies.Add(45690);
+                C.SelectedCurrencies.Add(48146);
+            }
             if (C.Version == "0.0.0") return;
-            if (C.Version == "1.2.0" && GetVersion() == "1.2.1") return;
             if (LowerVersionThan(GetVersion()))
             {
                 P.configWizard.IsOpen = true;